Split tests into multiple files
